Finish frontend
how website uses javascript to create transactions
and have wallet sign them.
- Use cast to interact with function selectors

Found something strange with the index.js and ABI
in constants.js where there are unnecessary
parameters for the fund function that are not
there on the contract. Tested and was able to
deploy a contract without fallback or receive,
changed the ABI and index.js and was able to
deposit using the correct function (matches
function selector). Opened up a discussion post
about it <https://github.com/Cyfrin/foundry-full-course-cu/discussions/2721>